Implementing Activity#getCurrentFocus and View#findFocus.
A View will return itself if it's in focus and a ViewGroup will
recursively look through its children for the first element that has
focus. FragmentActivity will search through its
fragments. Activity#getCurrentFocus calls findFocus on the contentView.

Each view keeps track of its own focus. When a view requests focus no
attempt is made of clearing focus from other views. Because of this,
ShadowActivty#clearFocus was added  to clear focus for all the View children.

Change-Id: If156cd79ca74d2736ffc404beba34525fe3ac9f7