Fix Wallet robolectric tests, Part 1
The build server java compiler builds robolectric tests in such
a way that custom shadows are added to class loader after the
test classes are loaded. This causes some custom shadows to be
ignored when the real classes are referenced in a test class.
This and the followup CL fixed that (hopefully) by moving the
addition of custom shadows forward to the constructor of
WalletRoboTestRunner.

Bug:10071558
Change-Id: I76745fdfcc08bdc198de8d90cf15dbabbbada7e0