Fix Robolectric ShadowSpannableStringBuilder
ShadowSpannableStringBuilder.append() was dropping non-String arguments
passed to it. This CL converts the argument to string before appending
it. This workaround allows tests to at least verify the stringified
version of the text. A real fix would be to properly implement
ShadowSpannableStringBuilder to support all CharSequence types and to
handle spans.

Change-Id: I26db6998f2389abb738a3edcb3dc0398ca4d3425