Fix resource loading issue.
Once robolectric tests are run with PanoSetupLib resources it
crashes with a NumberFormatException. This is caused by the
resource loader having the assumption that resources that
are referenced must be of the same type.

Such as:
<dimen name="foo">10dp</dimen>
<dimen name="bar">@dimen/foo</dimen>

However, the following case crashes (and is valid):
<integer name="foo">30dp</integer>
<dimen name="bar">@integer/foo</dimen>

This change explicitly allowed dimen resources to reference dimen
resources and integer resources.

Probably wouldn't need this change if we moved to v2 which uses
AOSP code for resource loading, but haven't had time to work on
the migration.

Change-Id: I726f69bd5740f31b7d9f1ae863a8792d764f3ee3