Switch ShadowBundle to use a TreeMap in robo-v1
Many Wallet tests assert equality of Parcelables by writing them
to Parcels and comparing the Parcels. This assumes a one-to-one
mapping between Parcelables and their Parcel representations.
The assumption holds for most Parcelables we come across, but breaks
down for Bundle. The Parcel representation of Bundle depends on the
order of the items in the Bundle. Previously ShadowBundle was backed
by an unordered HashMap. The order of items in the map can be affected
by the order in which the items are inserted. This change fixes that
by switching to an ordered TreeMap.

Change-Id: I5b6c3a6c4d6120176864f091fd59b2c7e8d94a69