Switch usages of legacy-test to android.test.base
The legacy-test library has been deprecated in favour of the
android.test.base library which is part of the SDK.

The following change descriptions were generated automatically and so
may be a little repetitive. They are provided to give the reviewer
enough information to check the comments match what has actually been
changed and check the reasoning behind the changes.

* Android.mk
    Replaced 'legacy-test' with 'android.test.base' in
    LOCAL_STATIC_JAVA_LIBRARIES because 'robolectric_android-all-stub'
    statically included legacy-test

Bug: 30188076
Test: make checkbuild
Change-Id: Ica12e65487ffe37c929f87f9c3a2f683458bf541